// src/app/api/models/workflow-form.ts
/* tslint:disable */
/* eslint-disable */
export interface WorkflowForm {
  createdAt?: string;
  defaultValue?: null | string;
  fieldName: string;
  fieldOrder?: number;
  id?: number;
  inputType?: null | string;
  isRequired?: boolean;
  options?: null | string;
  placeholder?: null | string;
  status?: null | string;
  stepNumber: number;
  updatedAt?: string;
  validation?: null | string;
  workflowId: number;
}